export proc-parser, bump for immediate release
