+++ packages/module/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@patternfly/quickstarts",
-  "version": "2.2.2",
+  "version": "2.2.3",
   "description": "PatternFly quick starts",
   "files": [
     "dist"
